requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
backports.cached-property==1.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
graphql-core==3.1.7
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.27.1
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
strawberry-graphql==0.96.0
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
